refactor(ChessmenInitializer): init method static으로 변경
